Update README.md
Improved the readability of the README file by reformatting various sections. 

Reformatted the installation steps to be clearer about what the user must do, and put where, in order to make FullCalendar display on the screen.

Merged the Usage section with the introductory paragraph.

Added Using FullCalendar section to link to helpful resources.